comments and white space cleanup in rust land
